+++ NEWS.md
@@ -1,3 +1,8 @@
+# lavaan.printer 0.0.0.9001
+
+- Add `est_funs_args`, `header_funs_args`,
+  and `footer_funs_args`. (0.0.0.9001)
+
 # lavaan.printer 0.0.0.9000
 
-* Initial CRAN submission.
+- Initial commit.